refactor: extract bootverify flow-boundary wiring checks
